Incremental adoption path for Webpacker packs with JSX, ES6, increasing linting strictness
